class refactor

+++ metchart/aggregator/dwd_icon.py
@@ -2,24 +2,119 @@
 
 import requests
 import datetime
-import pytz
 import requests
 import os
+import itertools
+
+import bz2
 
 from multiprocessing import cpu_count
 from multiprocessing.pool import ThreadPool
 
-import subprocess
-
 import xarray as xr
 
 from .. import misc
+from ..aggregator import Aggregator, Variable, Dimension
+
+from typing import Literal, Union
 
 BASE='https://opendata.dwd.de/weather/nwp'
 
+# NOTE: open_mfdataset() to open multiple gribs
+
+class IconAggregator(Aggregator):
+    PROVIDES = [
+        Variable.TEMPERATURE_3D,
+        Variable.U_3D,
+        Variable.V_3D,
+        Variable.U_SURFACE,
+        Variable.V_SURFACE,
+    ]
+
+    _MAPPING = {
+            Variable.TEMPERATURE_3D: {'path': 't', 'data':'t', 'plev': True},
+            Variable.U_3D: {'path': 'u', 'data':'u', 'plev': True},
+            Variable.V_3D: {'path': 'v', 'data':'v', 'plev': True},
+            Variable.U_SURFACE:{'path': 'u_10m', 'data':'u10', 'plev': False},
+            Variable.V_SURFACE:{'path': 'v_10m', 'data':'v10', 'plev': False},
+            Variable.GUST_SURFACE:{'path': 'vmax_10m', 'data':'fg10', 'plev': False},
+    }
+
+    _MODEL_SUFFIX = {
+            'icon': 'global',
+            'icon-eu': 'europe',
+            'icon-d2': 'germany'
+    }
+
+    def _init(self):
+        self._dataset = None
+
+    def _load_config(self, model: Literal['icon', 'icon-eu', 'icon-d2'],
+                     name: str, pressure_levels: list[int], steps: list[int],
+                     description: Union[None,str] = None) -> None:
+        self._name = name
+        self._description = description
+        self._model = model
+        self._levels = pressure_levels
+        self._steps = steps
+
+        self._download_dir = os.path.join(self._cache_dir, self._name)
+        misc.create_output_dir(self._download_dir)
+
+        self._caps_in_filename = False if model == 'icon-d2' else True
+
+    def _aggregate(self) -> None:
+        run, date = get_current_run()
+        filelist = self._list_needed_files(run, date)
+
+        for _ in ThreadPool(cpu_count()).map(download_url, filelist):
+            pass
+
+        _, gribs = zip(*filelist)
+        self._dataset = xr.open_mfdataset(gribs, engine='cfgrib')
+
+        # TODO is this needed still?
+        if self._description is not None:
+            self._dataset.attrs['_description'] = self._description
+
+    def _list_needed_files(self, run, date) -> list:
+        filelist = []
+
+        for step, level, var in itertools.product(
+                [f'{s:03d}' for s in self._steps],
+                self._levels,
+                [v for v in self._needed_variables if self._MAPPING[v]['plev']] ):
+            v = self._MAPPING[var]['path']
+            v_caps = v.upper() if self._caps_in_filename else v
+            filename = '{}_{}_regular-lat-lon_pressure-level_{}{}_{}_{}_{}.grib2'.format(
+                       self._model, self._MODEL_SUFFIX[self._model],
+                       date, run, step, level, v_caps )
+            URL = f'{BASE}/{self._model}/grib/{run}/{v}/{filename}.bz2'
+
+            filelist.append((URL, os.path.join(self._download_dir, filename)))
+
+        for step, var in itertools.product(
+                [f'{s:03d}' for s in self._steps],
+                [v for v in self._needed_variables if not self._MAPPING[v]['plev']] ):
+            v = self._MAPPING[var]['path']
+            v_caps = v.upper() if self._caps_in_filename else v
+            filename = '{}_{}_regular-lat-lon_single-level_{}{}_{}_{}.grib2'.format(
+                       self._model, self._MODEL_SUFFIX[self._model],
+                       date, run, step, v_caps )
+            URL = f'{BASE}/{self._model}/grib/{run}/{v}/{filename}.bz2'
+
+            filelist.append((URL, os.path.join(self._download_dir, filename)))
+
+        return filelist
+
+    def _query_data(self, var: Variable, query: list[tuple[Variable,object]]) -> xr.DataArray:
+        return xr.DataArray()
+
+    def _query_dimensions(self, var: Variable) -> list[Dimension]:
+        return []
+
 def get_current_run():
     # we allow up to 3h of slack for DWD to upload the latest run
-    tz = pytz.timezone('UTC')
     now = datetime.datetime.now(datetime.timezone.utc)
     corrected = now - datetime.timedelta(hours=3)
 
@@ -29,65 +124,17 @@ def get_current_run():
 
 def download_url(args):
     url, dest = args
+
+    if os.path.exists(dest):
+        return
+
     r = requests.get(url)
     try:
         with open(dest, 'wb') as f:
-            f.write(r.content)
+            f.write(bz2.decompress(r.content))
         print(f'Downloaded {dest}')
     except Exception as e:
         print(f'Failed to download {dest}:\n', e)
-
-def unpack_bz2(dest):
-    res = subprocess.run(['bzip2', '-df', dest])
-    if res.returncode != 0:
-        print(f'There was an error unpacking {dest}:', res.stderr)
-
-def download_dwd_gribs(
-        date, run, target, output, model, steps, model_long,
-        pressure_level_parameters, parameter_caps_in_filename,
-        single_level_parameters, pressure_levels
-):
-    misc.create_output_dir(output)
-
-    to_download = []
-
-    for step in steps:
-        step_str = f'{step:03d}'
-
-        for parameter in pressure_level_parameters:
-            parameter2 = parameter.upper() if parameter_caps_in_filename else parameter
-
-            for level in pressure_levels:
-                filename = f'{model_long}_regular-lat-lon_pressure-level_{date}{run}_{step_str}_{level}_{parameter2}.grib2.bz2'
-                URL = f'{BASE}/{model}/grib/{run}/{parameter}/{filename}'
-
-                to_download.append((URL, os.path.join(output, filename)))
-
-        for parameter in single_level_parameters:
-            parameter2 = parameter.upper() if parameter_caps_in_filename else parameter
-            filename = f'{model_long}_regular-lat-lon_single-level_{date}{run}_{step_str}_{parameter2}.grib2.bz2'
-            URL = f'{BASE}/{model}/grib/{run}/{parameter}/{filename}'
-
-            to_download.append((URL, os.path.join(output, filename)))
-
-
-    for _ in ThreadPool(cpu_count()).imap_unordered(download_url, to_download):
-        pass
-
-    print('Done Downloading. Uncompressing...')
-
-    for _ in ThreadPool(cpu_count()).imap_unordered(unpack_bz2, [dest for _, dest in to_download]):
-        pass
-
-    downloaded_gribs = [dest.removesuffix('.bz2') for _, dest in to_download]
-
-    res = subprocess.run(['grib_copy'] + downloaded_gribs + [target])
-    if res.returncode != 0:
-        print('grib_copy failed with: ', res.stderr)
-
-    res = subprocess.run(['rm', '-f'] + downloaded_gribs)
-    if res.returncode != 0:
-        print('rm failed with: ', res.stderr)
 
 def clean_output_dir(directory, target):
     to_delete = [f for f  in os.listdir(directory) if os.path.isfile(os.path.join(directory, f))]
@@ -96,54 +143,3 @@ def clean_output_dir(directory, target):
 
     for f in to_delete:
         os.unlink(os.path.join(directory, f))
-
-def load_data(name, output, description = None, clean = False, force_filename = None, **kwargs):
-    target = force_filename
-
-    if target is None:
-        run, date = get_current_run()
-        filename = f'{name}_{date}_{run}.grib2'
-        target = os.path.join(output, filename)
-
-        if not os.path.exists(target):
-            download_dwd_gribs(date, run, target, output, **kwargs)
-        else:
-            print(f'{target} already exists. Using the cached version.')
-
-        if clean:
-            clean_output_dir(output, filename)
-
-    # we drop heightAboveGround to allow 2m and 10m values to be merged down to one dataset
-    ds = xr.load_dataset(target, engine='cfgrib', drop_variables='heightAboveGround')
-
-    if description is not None:
-        ds.attrs['_description'] = description
-    return ds
-
-
-debug_config = {
-    'output':'dwd_icon-eu',
-    'model':'icon-eu',
-    'model_long':'icon-eu_europe',
-    'clean': True,
-    'parameter_caps_in_filename':True,
-    'pressure_level_parameters': [
-        't',
-        'relhum',
-        'u',
-        'v',
-        'fi',
-        'clc'
-    ],
-    'single_level_parameters': [
-        'pmsl',
-        't_2m',
-        'relhum_2m'
-    ],
-    'pressure_levels':[ 1000, 950, 925, 900, 875, 850, 825, 800, 775, 700, 600, 500, 400, 300, 250, 200, 150, 100 ],
-    'steps':[0, 3, 6, 9, 12, 15, 18, 21, 24, 27, 30, 33, 36, 39, 42, 45, 48]
-}
-
-if __name__ == '__main__':
-    load_data('test_icon_eu', **debug_config)
-
